Fix iOS compilation

+++ Xcode8Project/Dungeons of Noudar/Data/Assets/fragment.glsl
@@ -1,5 +1,4 @@
-precision highp float;
-
+precision mediump float;
 varying vec2 vTextureCoords;
 varying vec4 vColour;
 varying float distance;
@@ -9,15 +8,15 @@ uniform vec4 uFade;
 uniform vec4 uFog;
 
 void main() {
-	
-	float fade = 0.75 + ((1.0 - distance) / 4.0);
-	gl_FragColor = texture2D( sTexture, vTextureCoords ) * uMod * vec4( fade, fade, fade, 1.0 );
-	
-	if ( gl_FragColor.a < 0.5 ) {
-		discard;
-	}
-	
-	if ( uFade.a >= 0.1 ) {
-		gl_FragColor = gl_FragColor * vec4( uFade.xyz, 1.0 );
-	}
+
+    float fade = 0.75 + ((1.0 - distance) / 4.0);
+    gl_FragColor = texture2D( sTexture, vTextureCoords ) * uMod * vec4( fade, fade, fade, 1.0 );
+
+    if ( gl_FragColor.a < 0.5 ) {
+        discard;
+    }
+
+    if ( uFade.a >= 0.1 ) {
+        gl_FragColor = gl_FragColor * vec4( uFade.xyz, 1.0 );
+    }
 }
